requirements.txt
certifi==2025.6.15
langchain==0.3.26
langchain_community==0.3.26
langchain_core==0.3.66
langchain_groq==0.3.4
langchain_huggingface==0.3.0
langchain_ollama==0.3.3
langchain_openai==0.3.26
langchain_qdrant==0.2.0
pdf2image==1.17.0
Pillow==11.2.1
pydantic==2.11.7
pydub==0.25.1
PyPDF2==3.0.1
pytesseract==0.3.13
python-dotenv==1.1.1
qdrant_client==1.14.3
Requests==2.32.4
streamlit==1.45.0
streamlit_mic_recorder==0.0.8
urllib3>=1.26.14,<3
torch==2.3.0
sentence-transformers==2.6.1
numpy==1.26.4
tomli==2.0.1
